service-workers/service-worker/registration-service-worker-attributes.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';
promise_test(async t => {
  const SCOPE = 'resources/scope/installing-waiting-active-after-registration';
  const WORKER_URL = 'resources/empty-worker.js';
  const expected_url = normalizeURL(WORKER_URL);

  const registration =
      await service_worker_unregister_and_register(t, WORKER_URL, SCOPE);
  const newest_worker = await registration.installing;
  
  t.add_cleanup(async() => {
    if (registration)
      await registration.unregister();
  });

  assert_equals(registration.installing.scriptURL, expected_url,
                'installing before updatefound');
  assert_equals(registration.waiting, null, 'waiting before updatefound');
  assert_equals(registration.active, null, 'active before updatefound');

  await wait_for_update(t, registration);
  assert_equals(registration.installing, newest_worker,
                'installing after updatefound');
  assert_equals(registration.waiting, null, 'waiting after updatefound');
  assert_equals(registration.active, null, 'active after updatefound');

  await wait_for_state(t, registration.installing, 'installed');
  assert_equals(registration.installing, null, 'installing after installed');
  assert_equals(registration.waiting, newest_worker,
                'waiting after installed');
  assert_equals(registration.active, null, 'active after installed');

  await wait_for_state(t, registration.waiting, 'activated');
  assert_equals(registration.installing, null, 'installing after activated');
  assert_equals(registration.waiting, null, 'waiting after activated');
  assert_equals(registration.active, newest_worker, 'active after activated');
  await Promise.all([
        wait_for_state(t, registration.active, 'redundant'),
        registration.unregister()
  ]);
  assert_equals(registration.installing, null, 'installing after redundant');
  assert_equals(registration.waiting, null, 'waiting after redundant');
  // According to spec, Clear Registration runs Update State which is
  // immediately followed by setting active to null, which means by the
  // time the event loop turns and the Promise for statechange is
  // resolved, this will be gone.
  assert_equals(registration.active, null,
                'active should be null after redundant');
});
</script>
