LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/registration-service-worker-attributes.https-expected.txt


FAIL installing/waiting/active after registration assert_equals: active should be null after redundant expected null but got object "[object ServiceWorker]"
